move `TreeItem` into its own file add missing import to `styles.css` remove redecalarations
