+++ packages/kiwi-react/src/bricks/styles.css
@@ -30,5 +30,6 @@
 @import "./TextBox.css" layer(itwinui.components);
 @import "./Tooltip.css" layer(itwinui.components);
 @import "./Tree.css" layer(itwinui.components);
+@import "./TreeItem.css" layer(itwinui.components);
 
 @import "./Field.css" layer(itwinui.components);
